requirements.txt
discord.py
pyyaml~=5.3.1
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
asyncio~=3.4.3
python-dateutil~=2.8.1
pytz~=2020.1
google~=3.0.0
beautifulsoup4